Added Rigid Body to Knife and updated gravity settings in game
After adding the new knife sprite I was still having trouble with the Gravity settings in the game. So I added a new rigid body to the knife sprite and then update the gravity settings in the game to y=-9.81. Now the knife will fall from the ceiling or top of court.