Changed Knife property to is-trigger and updated tag on player and enemy to player
Change collider on knife to is Trigger so the knife will pass through the players and not change their place on the court when they are hit. I also updated the tags on the players so we can detect when the knife only hits a player and not the balls.